refactor: eliminate brightray::BrowserClient
