+++ brightray/common/main_delegate.cc
@@ -9,7 +9,6 @@
 #include "base/command_line.h"
 #include "base/mac/bundle_locations.h"
 #include "base/path_service.h"
-#include "brightray/browser/browser_client.h"
 #include "content/public/common/content_switches.h"
 #include "electron/buildflags/buildflags.h"
 #include "services/service_manager/embedder/switches.h"
@@ -95,15 +94,6 @@ void MainDelegate::PreSandboxStartup() {
     std::string locale = cmd.GetSwitchValueASCII(switches::kLang);
     LoadResourceBundle(locale);
   }
-}
-
-content::ContentBrowserClient* MainDelegate::CreateContentBrowserClient() {
-  browser_client_ = CreateBrowserClient();
-  return browser_client_.get();
-}
-
-std::unique_ptr<BrowserClient> MainDelegate::CreateBrowserClient() {
-  return std::unique_ptr<BrowserClient>(new BrowserClient);
 }
 
 }  // namespace brightray
