+++ brightray/common/main_delegate.h
@@ -21,8 +21,6 @@ class ResourceBundle;
 
 namespace brightray {
 
-class BrowserClient;
-
 void LoadResourceBundle(const std::string& locale);
 void LoadCommonResources();
 
@@ -32,10 +30,6 @@ class MainDelegate : public content::ContentMainDelegate {
   ~MainDelegate() override;
 
  protected:
-  // Subclasses can override this to provide their own BrowserClient
-  // implementation.
-  virtual std::unique_ptr<BrowserClient> CreateBrowserClient();
-
 #if defined(OS_MACOSX)
   // Subclasses can override this to custom the paths of child process and
   // framework bundle.
@@ -47,10 +41,6 @@ class MainDelegate : public content::ContentMainDelegate {
   void PreSandboxStartup() override;
 
  private:
-  content::ContentBrowserClient* CreateContentBrowserClient() override;
-
-  std::unique_ptr<BrowserClient> browser_client_;
-
   DISALLOW_COPY_AND_ASSIGN(MainDelegate);
 };
